+++ BLE_Hardware_ESP32/gatt_server/.vscode/c_cpp_properties.json
@@ -2,7 +2,7 @@
     "configurations": [
         {
             "name": "ESP-IDF",
-            "compilerPath": "${config:idf.toolsPathWin}\\tools\\xtensa-esp-elf\\esp-13.2.0_20240530\\xtensa-esp-elf\\bin\\xtensa-esp32-elf-gcc.exe",
+            "compilerPath": "${config:idf.toolsPathWin}\\tools\\xtensa-esp-elf\\esp-14.2.0_20241119\\xtensa-esp-elf\\bin\\xtensa-esp32-elf-gcc.exe",
             "compileCommands": "${config:idf.buildPath}/compile_commands.json",
             "includePath": [
                 "${config:idf.espIdfPath}/components/**",
