+++ BLE_Hardware_ESP32/gatt_server/.vscode/settings.json
@@ -1,11 +1,8 @@
 {
-  "C_Cpp.intelliSenseEngine": "default",
-  "idf.flashType": "UART",
-  "idf.portWin": "COM3",
-  "idf.adapterTargetName": "esp32",
-  "files.associations": {
-    "esp_log.h": "c",
-    "freertos.h": "c",
-    "stddef.h": "c"
-  }
+    "C_Cpp.intelliSenseEngine": "default",
+    "idf.espIdfPathWin": "C:\\Users\\annmon.james\\esp\\v5.4\\esp-idf",
+    "idf.toolsPathWin": "C:\\Users\\annmon.james\\.espressif",
+    "idf.pythonInstallPath": "C:\\Users\\annmon.james\\.espressif\\tools\\idf-python\\3.11.2\\python.exe",
+    "idf.flashType": "UART",
+    "idf.portWin": "COM3"
 }
