+++ BLE_Hardware_ESP32/gatt_server/main/Kconfig.projbuild
@@ -1,6 +1,6 @@
 menu "Example 'GATT SERVER' Config"
 
-    config SET_RAW_ADV_DATA
+    config EXAMPLE_SET_RAW_ADV_DATA
         bool "Use raw data for advertising packets and scan response data"
         help
             If this config item is set, raw binary data will be used to generate advertising & scan response data.
@@ -10,5 +10,17 @@ menu "Example 'GATT SERVER' Config"
             If this config item is unset, advertising & scan response data is provided via a higher-level
             esp_ble_adv_data_t structure. The lower layer will generate the BLE packets. This option has higher
             overhead at runtime.
+
+    config EXAMPLE_CI_ID
+        int
+        default 70
+        help
+            This config the example id for CI test. Only for internal used.
+
+    config EXAMPLE_CI_PIPELINE_ID
+        int "The pipeline id for CI test"
+        default 0
+        help
+            This config the pipeline id for CI test. Only for internal used.
 
 endmenu
