+++ BLE_Hardware_ESP32/gatt_server/sdkconfig.ci.esp32c2_xtal26m
@@ -1,0 +1,4 @@
+CONFIG_EXAMPLE_CI_ID=1
+CONFIG_IDF_TARGET="esp32c2"
+CONFIG_XTAL_FREQ_26=y
+CONFIG_EXAMPLE_CI_PIPELINE_ID=${CI_PIPELINE_ID}
