+++ BLE_Hardware_ESP32/gatt_server/sdkconfig.ci.name
@@ -1,0 +1,3 @@
+
+CONFIG_EXAMPLE_CI_ID=1
+CONFIG_EXAMPLE_CI_PIPELINE_ID=${CI_PIPELINE_ID}
