+++ BLE_Hardware_ESP32/gatt_server/sdkconfig.defaults.esp32c2
@@ -6,3 +6,6 @@ CONFIG_BT_ENABLED=y
 CONFIG_BT_BLE_42_FEATURES_SUPPORTED=y
 # CONFIG_BT_LE_50_FEATURE_SUPPORT is not set
 CONFIG_BT_LE_HCI_EVT_BUF_SIZE=257
+# XTAL Freq Config
+CONFIG_XTAL_FREQ_26=y
+CONFIG_XTAL_FREQ=26
